TheFaceOfLynbrook/requirements.txt
aioflask==0.4.0
altgraph==0.17
anyio==3.6.1
appdirs==1.4.4
appnope==0.1.3
asgiref==3.6.0
asttokens==2.0.7
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
cycler==0.10.0
Cython==0.29.21
debugpy==1.6.2
decorator==5.1.1
deta==1.1.0
dill==0.3.3
distlib==0.3.1
emoji==1.7.0
entrypoints==0.4
executing==0.9.1
fastapi==0.79.0
filelock==3.0.12
Flask==2.2.2
greenlet==2.0.2
greenletio==0.9.0
h11==0.13.0
idna==3.3
importlib-metadata==6.0.0
ipykernel==6.15.1
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.0.0
jupyter-client==7.3.4
jupyter-core==4.11.1
kiwisolver==1.3.1
macholib==1.14
MarkupSafe==2.1.2
matplotlib==3.3.4
matplotlib-inline==0.1.3
multiprocess==0.70.11.1
nest-asyncio==1.5.5
numpy==1.19.5
pandas==1.2.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.9.2
Pygments==2.12.0
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.2
pyparsing==2.4.7
PyQt3D==5.15.5
PyQt5==5.15.7
PyQt5-sip==12.11.0
PyQtChart==5.15.6
PyQtDataVisualization==5.15.5
PyQtNetworkAuth==5.15.5
PyQtPurchasing==5.15.5
PyQtWebEngine==5.15.6
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2020.5
pyzmq==23.2.0
requests==2.27.1
scikit-learn==0.24.2
scipy==1.6.0
seaborn==0.11.1
sealion==4.1.3
six==1.15.0
sklearn==0.0
sniffio==1.2.0
stack-data==0.3.0
starlette==0.19.1
threadpoolctl==2.2.0
tornado==6.2
tqdm==4.56.0
traitlets==5.3.0
typing-extensions==3.10.0.0
urllib3==1.26.9
uvicorn==0.18.2
virtualenv==20.2.2
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.12.0
